+++ adsft/tests/test_tasks.py
@@ -14,7 +14,7 @@ class TestWorkers(unittest.TestCase):
         unittest.TestCase.setUp(self)
         self.proj_home = os.path.join(os.path.dirname(__file__), '../..')
         self._app = tasks.app
-        self.app = app.create_app('test',
+        self.app = app.ADSFulltextCelery('test',
             {
             'SQLALCHEMY_URL': 'sqlite:///',
             'SQLALCHEMY_ECHO': False
